fix: Cast x value in reddit to int
Fixes macOS builds? I hope?

+++ natives/image/reddit.cc
@@ -46,9 +46,9 @@ CmdOutput esmb::Image::Reddit(const string &type, string &outType, const char *b
                                         ->set("font", "Roboto 62")
                                         ->set("fontfile", (basePath + "assets/fonts/reddit.ttf").c_str()));
 
-  VImage composited =
-    tmpl.composite2(textImage, VIPS_BLEND_MODE_OVER,
-                    VImage::option()->set("x", 64)->set("y", (tmpl.height() - baseline) - 64 + textImage.yoffset()));
+  VImage composited = tmpl.composite2(
+    textImage, VIPS_BLEND_MODE_OVER,
+    VImage::option()->set("x", (int)64)->set("y", (tmpl.height() - baseline) - 64 + textImage.yoffset()));
   VImage watermark = composited.resize((double)width / (double)composited.width());
 
   vector<VImage> img;
